Moved some utility files back
