+++ src/act/alexandria/tests/CMakeLists.txt
@@ -39,6 +39,7 @@ gmx_add_unit_test(
     ${testname}
     ${exename}
     # files with code for tests
+    actmol_util.cpp
     allbondedsTest.cpp
     confighandlerTest.cpp
     devcomputerTest.cpp
