+++ src/act/alexandria/tests/forcecomputerTest.cpp
@@ -42,10 +42,11 @@
 
 #include <gtest/gtest.h>
 
+#include "actmol_util.h"
+
 #include "act/import/import.h"
 #include "act/forcefield/forcefield_utils.h"
 #include "act/topology/actmol.h"
-#include "act/topology/actmol_util.h"
 #include "gromacs/utility/fatalerror.h"
 
 #include "testutils/refdata.h"
